[#71]/refactor ProfileImageForm 컴포넌트 분리
