+++ web/src/components/mypage/ProfileEditForm.tsx
@@ -1,99 +1,19 @@
 /* eslint-disable no-alert */
 /* eslint-disable no-console */
 import { changeProfileImage, nicknameCheck } from 'api/userAPI';
-import { ColorizeIcon, X16BigIcon } from 'assets/icons';
 import SimpleButton from 'components/common/SimpleButton';
 import SimpleInput from 'components/common/SimpleInput';
 import SmallBlackLabel from 'components/common/SmallBlackLabel';
-import useToggle from 'hooks/common/useToggle';
 import React, { ReactElement, useState } from 'react';
 import { useRecoilState } from 'recoil';
 import { userState } from 'recoil/atoms/userState';
 import styled from 'styled-components';
 import { DEFAULT_IMAGE_FILE_NAME } from 'utils/const';
-import ProfileColorPalette from './ProfileColorPalette';
+import ProfileImageForm from './ProfileImageForm';
 
 const ProfileEditFormWrapper = styled.div`
   padding-top: 24px;
   color: ${(props) => props.theme.color.grayDarkest};
-`;
-
-const ImgFormRow = styled.div`
-  margin-bottom: 45px;
-  display: flex;
-`;
-
-const FormLabel = styled(SmallBlackLabel)`
-  padding-top: 20px;
-`;
-
-const ProfileImageBox = styled.div`
-  width: 72px;
-  height: 72px;
-  position: relative;
-`;
-
-const ProfileImage = styled.img`
-  width: 100%;
-  height: 100%;
-  border-radius: 40px;
-  background-color: ${(props) => props.theme.color.grayLight};
-`;
-
-const ProfileColorsButton = styled(ColorizeIcon)`
-  position: absolute;
-  bottom: -2px;
-  right: -11px;
-  cursor: pointer;
-`;
-
-const UploadContent = styled.div`
-  margin-left: 27px;
-`;
-
-const UploadRow = styled.div`
-  display: flex;
-  align-items: center;
-  font-size: 14px;
-  color: ${(props) => props.theme.color.grayDark};
-  font-weight: 300;
-  &:first-child {
-    margin-bottom: 8px;
-    padding-top: 14px;
-  }
-  &:last-child {
-    font-size: 12px;
-  }
-`;
-
-const UploadButton = styled.label`
-  width: 75px;
-  height: 31px;
-  border-radius: 6px;
-  font-weight: 400;
-  line-height: 1.5;
-  font-size: 14px;
-  background-color: #ffffff;
-  color: #323232;
-  border: 1px solid #aaaaaa;
-  margin-right: 16px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-`;
-
-const DeleteButton = styled(X16BigIcon)`
-  cursor: pointer;
-`;
-
-const FileInputStyled = styled.input`
-  display: none; ;
-`;
-
-const UploadFileName = styled.span`
-  line-height: 1.5;
-  margin-right: 4px;
 `;
 
 const NicknameFormRow = styled.div`
@@ -127,7 +47,6 @@ const SaveButton = styled(SimpleButton)`
 `;
 
 function ProfileEditForm(): ReactElement {
-  const [isPaletteOpen, onTogglePaletteOpen] = useToggle();
   const [user, setUser] = useRecoilState(userState);
   const [form, setForm] = useState({
     profileImage: user.imageUrl,
@@ -183,37 +102,12 @@ function ProfileEditForm(): ReactElement {
 
   return (
     <ProfileEditFormWrapper>
-      <ImgFormRow>
-        <FormLabel width="297px" label="프로필 이미지" />
-        <ProfileImageBox>
-          <ProfileImage src={profileImage} />
-          <ProfileColorsButton onClick={onTogglePaletteOpen} />
-          {isPaletteOpen && (
-            <ProfileColorPalette
-              isOpen={isPaletteOpen}
-              onToggleOpen={onTogglePaletteOpen}
-              onChangeProfileImage={onChangeProfileImage}
-            />
-          )}
-        </ProfileImageBox>
-
-        <UploadContent>
-          <UploadRow>
-            <UploadButton htmlFor="profile-image-upload">
-              파일 선택
-            </UploadButton>
-            <FileInputStyled
-              type="file"
-              id="profile-image-upload"
-              // eslint-disable-next-line no-console
-              onChange={onImageUpload}
-            />
-            <UploadFileName>{imageFileName}</UploadFileName>
-            <DeleteButton onClick={onDeleteImage} />
-          </UploadRow>
-          <UploadRow>최대 10MB의 이미지 파일</UploadRow>
-        </UploadContent>
-      </ImgFormRow>
+      <ProfileImageForm
+        form={form}
+        onChangeProfileImage={onChangeProfileImage}
+        onDeleteImage={onDeleteImage}
+        onImageUpload={onImageUpload}
+      />
 
       <NicknameFormRow>
         <NicknameFormLabel width="297px" label="닉네임" />
